+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ##### 0.10.1 - xx July 2014
 
 ###### Backwards compatible bug fixes
-- #90 - DS.create isn't added to completedQueries
-- #91 - dist/angular-data(.min).js doesn't end with a semicolon
-- #95 - observe-js outdated
+- #90 - DS.create isn't added to completedQueries (`DS.create` now adds a completed query entry)
+- #91 - dist/angular-data(.min).js doesn't end with a semicolon (upgraded Browserify)
+- #94 - Resource object name/class inconsistency (added `useClass` option to `DS.defineResource`)
+- #95 - observe-js outdated (Upgraded observe-js.js an refactored to new API)
 
 ##### 0.10.0 - 18 July 2014